requirements.txt
fastapi
uvicorn
aiogram==3.3.0
aiohttp
openai
python-dotenv
aiocryptopay==1.0.1
pydantic  # для FastAPI

